4857: add Language/UserRole form fields
