+++ src/client/components/Form/FormFields/FormFields.tsx
@@ -1,13 +1,17 @@
 import React from 'react'
 
-import { FormFieldType } from '../types'
-import PermissionsField from './PermissionsField'
-import SelectField from './SelectField'
-import TextField from './TextField'
-import { FieldProps } from './types'
+import LanguageField from 'client/components/Form/FormFields/LanguageField'
+import PermissionsField from 'client/components/Form/FormFields/PermissionsField'
+import SelectField from 'client/components/Form/FormFields/SelectField'
+import TextField from 'client/components/Form/FormFields/TextField'
+import { FieldProps } from 'client/components/Form/FormFields/types'
+import UserRoleField from 'client/components/Form/FormFields/UserRoleField'
+import { FormFieldType } from 'client/components/Form/types'
 
 export const FormFields: Record<FormFieldType, React.FC<FieldProps>> = {
-  [FormFieldType.text]: TextField,
-  [FormFieldType.select]: SelectField,
+  [FormFieldType.language]: LanguageField,
   [FormFieldType.permissions]: PermissionsField,
+  [FormFieldType.select]: SelectField,
+  [FormFieldType.text]: TextField,
+  [FormFieldType.userRole]: UserRoleField,
 }
